refactor : [리뷰사항반영] 도메인명을 `Comment`로 변경함에 따른 이름 수정
