+++ src/main/java/kr/codesqaud/cafe/controller/dto/CommentResponse.java
@@ -3,9 +3,9 @@ package kr.codesqaud.cafe.controller.dto;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-import kr.codesqaud.cafe.domain.articlecomment.Comment;
+import kr.codesqaud.cafe.domain.comment.Comment;
 
-public class ArticleCommentResponse {
+public class CommentResponse {
 
 	private static final DateTimeFormatter DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
 
@@ -14,15 +14,15 @@ public class ArticleCommentResponse {
 	private final String content;
 	private final String createdAt;
 
-	public ArticleCommentResponse(Long id, String writer, String content, LocalDateTime createdAt) {
+	public CommentResponse(Long id, String writer, String content, LocalDateTime createdAt) {
 		this.id = id;
 		this.writer = writer;
 		this.content = content;
 		this.createdAt = DATE_TIME_FORMATTER.format(createdAt);
 	}
 
-	public static ArticleCommentResponse from(final Comment comment) {
-		return new ArticleCommentResponse(
+	public static CommentResponse from(final Comment comment) {
+		return new CommentResponse(
 			comment.getId(),
 			comment.getWriter(),
 			comment.getContent(),
