+++ src/main/resources/templates/fragments/header.html
@@ -9,5 +9,6 @@
     <!--[if lt IE 9]>
     <script src="//html5shim.googlecode.com/svn/trunk/html5.js"></script>
     <![endif]-->
+    <script src="https://code.jquery.com/jquery-3.4.1.js"></script>
     <link href="/css/styles.css" rel="stylesheet">
 </head>
