+++ src/main/resources/templates/qna/show.html
@@ -46,7 +46,7 @@
                 {{/article}}
                 <div class="qna-comment">
                     <div class="qna-comment-slipp">
-                        <p class="qna-comment-count"><strong>{{articleCommentCount}}</strong>개의 의견</p>
+                        <p class="qna-comment-count"><strong>{{commentCount}}</strong>개의 의견</p>
                         <div class="qna-comment-slipp-articles">
                             {{#comments}}
                             <article class="article" id="answer-{{id}}">
@@ -122,7 +122,7 @@
                     <a class="link-modify-article" href="/api/qna/updateAnswer/{3}">수정</a>
                 </li>
                 <li>
-                    <form class="delete-answer-form" action="/api/questions/{3}/answers/{4}" method="POST">
+                    <form class="delete-answer-form" action="/comments/{4}" method="POST">
                         <input type="hidden" name="_method" value="DELETE">
                         <button type="submit" class="delete-answer-button">삭제</button>
                     </form>
